+++ src/module/applications/sidebar/_module.mjs
@@ -1,1 +1,2 @@
+export * as apps from "./apps/_module.mjs";
 export * as tabs from "./tabs/_module.mjs";