+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Implemented a Reference page, which has additional room to specify a tooltip for the reference enricher. (#1423, #1425)
 - Added `system.artDescription` as a valid property for Image journal entry pages, providing rich alt text. (#1424)
 - Added new embed for treasure items with an option for including project info (defaults to false). (#1462)
+- Implemented a Table of Contents view for journal compendiums. (#1502)
 - Made the function to apply an NPC's free strike publicly accessible at `actor.system.performFreeStrike()`.
 
 ### Changed